Fix TypeError crash: numpy float32 rejected by ultralytics conf=
Real production crash the user hit uploading a video: TypeError:
'conf=0.2598804235458374' is of invalid type float32. Valid 'conf'
types are int (i.e. 'conf=0') or float (i.e. 'conf=0.5'), raised
inside the pothole detector's YOLO call during video processing.

Root cause: scene_aware_conf()'s luminance argument is always a
numpy float32 in production (the mean of a frame array), so its
arithmetic -- and the confidence value it returns -- inherited that
numpy dtype instead of being a native Python float. A newer
ultralytics release started strictly validating conf= as int/float
and rejecting numpy scalar subtypes. This function's return value
feeds directly into all 8 conf= call sites across
app/streamlit_app.py (video processing, image tab, live snapshot,
live stream, both main and pothole detectors), so every one of them
was affected, not just the path the user happened to hit first.

Fixed at the single source: scene_aware_conf() now explicitly
returns float(...), fixing every call site at once instead of
patching each individually.

Validation: reproduced the exact crash directly against the real
installed ultralytics version via ultralytics.cfg.check_cfg()
(confirmed numpy.float32 conf is rejected, a native float is
accepted) -- confirmed root cause, not a guess. Added
tests/test_enhance.py (7 tests -- there was zero prior coverage for
scene_aware_conf/scene_blend_weight, which is exactly why this went
unnoticed) including a regression test asserting type(result) is
float given a numpy.float32 luminance, the real production input
type. Live-tested the actual Video tab end-to-end
(test_video_tab_processes, test_video_cancel_stops_promptly), which
exercises the exact pothole-detector call path that crashed, since
app/pothole_best.pt is committed and loads in these tests too.

Full suite (48 tests) passes.

+++ src/enhance.py
@@ -69,7 +69,14 @@ def scene_aware_conf(base_conf, luminance, dark_thresh=0.35,
     between the two (dusk, shaded hillside roads).
     """
     weight = scene_blend_weight(luminance, dark_thresh, bright_thresh)
-    return min(base_conf + night_boost * weight, 0.95)
+    # luminance is normally a numpy float32 (mean of a uint8/float32
+    # frame array), which makes `weight` -- and so the result here --
+    # inherit that dtype through the arithmetic. Cast to a native
+    # Python float: newer ultralytics versions strictly validate
+    # conf= as int/float and reject numpy scalar subtypes, and every
+    # caller of this function feeds its result straight into a YOLO
+    # conf= argument.
+    return float(min(base_conf + night_boost * weight, 0.95))
 
 
 def correct_color_cast(frame_uint8, strength=0.6, lum_pctl=60, max_shift=18):
